+++ src/form/b-radio-button/b-radio-button.ts
@@ -6,7 +6,17 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
-import bCheckbox, { component, Value, FormValue } from 'form/b-checkbox/b-checkbox';
+import bCheckbox, {
+
+	component,
+	Value,
+	FormValue,
+	ValidatorsDecl,
+	ValidatorParams,
+	ValidatorResult
+
+} from 'form/b-checkbox/b-checkbox';
+
 export * from 'super/i-input/i-input';
 
 @component({flyweight: true})
@@ -16,11 +26,42 @@ export default class bRadioButton<
 	D extends object = Dictionary
 > extends bCheckbox<V, FV, D> {
 	/** @override */
-	protected async onClick(e: Event): Promise<void> {
-		if (this.mods.checked === 'true') {
-			return;
+	static validators: ValidatorsDecl = {
+		//#if runtime has iInput/validators
+
+		async required({msg, showMsg = true}: ValidatorParams): Promise<ValidatorResult<boolean>> {
+			const
+				value = await this.groupFormValue;
+
+			if (!value) {
+				this.setValidationMsg(this.getValidatorMsg(false, msg, t`Required field`), showMsg);
+				return false;
+			}
+
+			return true;
 		}
 
-		return super.onClick(e);
+		//#endif
+	};
+
+	/** @override */
+	protected async onClick(e: Event): Promise<void> {
+		await this.focus();
+
+		if (await this.check()) {
+			const
+				ctx = <any>this;
+
+			for (let els = await this.groupElements, i = 0; i < els.length; i++) {
+				const
+					el = els[i];
+
+				if (el !== ctx && this.isComponent(el, bRadioButton)) {
+					el.uncheck().catch(stderr);
+				}
+			}
+
+			this.emit('actionChange', true);
+		}
 	}
 }