+++ src/super/i-block/i-block.ts
@@ -13,7 +13,6 @@ import symbolGenerator from 'core/symbol';
 import Async, { AsyncOpts, ClearOptsId, WrappedFunction, ProxyCb } from 'core/async';
 import log, { LogMessageOpts } from 'core/log';
 
-import { GLOBAL } from 'core/env';
 import { EventEmitter2 as EventEmitter } from 'eventemitter2';
 
 //#if runtime has core/helpers
@@ -151,11 +150,8 @@ export {
 	system,
 	watch,
 	wait,
-	bindModTo,
 	mod,
-	removeMod,
-	elMod,
-	removeElMod
+	removeMod
 
 } from 'super/i-block/modules/decorators';
 
@@ -934,7 +930,7 @@ export default abstract class iBlock extends ComponentInterface<iBlock, iStaticP
 		replace: true
 	})
 
-	protected readonly l: typeof l = GLOBAL.l;
+	protected readonly l: typeof l = globalThis.l;
 
 	/**
 	 * Link to console API
@@ -967,7 +963,7 @@ export default abstract class iBlock extends ComponentInterface<iBlock, iStaticP
 		atom: true,
 		unique: true,
 		replace: true,
-		init: () => GLOBAL
+		init: () => globalThis
 	})
 
 	protected readonly global!: Window;
@@ -1730,7 +1726,7 @@ export default abstract class iBlock extends ComponentInterface<iBlock, iStaticP
 	 * @param obj
 	 * @param [constructor] - component constructor
 	 */
-	protected isComponent<T extends iBlock>(obj: unknown, constructor?: {new(): T}): obj is T {
+	protected isComponent<T extends iBlock>(obj: unknown, constructor?: {new(): T} | Function): obj is T {
 		return Boolean(obj && (<Dictionary>obj).instance instanceof (constructor || iBlock));
 	}